refactor(backend-game-domain): move game models
