+++ packages/backend/apps/game/backend-game-domain/src/games/domain/entities/ActiveGame.ts
@@ -1,4 +1,4 @@
-import { ActiveGameState } from './ActiveGameState';
+import { ActiveGameState } from '../valueObjects/ActiveGameState';
 import { BaseGame } from './BaseGame';
 
 export type ActiveGame = BaseGame<ActiveGameState>;
