+++ packages/backend/apps/game/backend-game-domain/src/games/domain/entities/NonStartedGame.ts
@@ -1,4 +1,4 @@
+import { NonStartedGameState } from '../valueObjects/NonStartedGameState';
 import { BaseGame } from './BaseGame';
-import { NonStartedGameState } from './NonStartedGameState';
 
 export type NonStartedGame = BaseGame<NonStartedGameState>;
