+++ packages/backend/apps/game/backend-game-domain/src/games/domain/fixtures/ActiveGameFixtures.ts
@@ -6,9 +6,9 @@ import { ReverseCard } from '../../../cards/domain/valueObjects/ReverseCard';
 import { SkipCard } from '../../../cards/domain/valueObjects/SkipCard';
 import { WildCard } from '../../../cards/domain/valueObjects/WildCard';
 import { WildDraw4Card } from '../../../cards/domain/valueObjects/WildDraw4Card';
-import { ActiveGame } from '../models/ActiveGame';
-import { GameDirection } from '../models/GameDirection';
-import { GameStatus } from '../models/GameStatus';
+import { ActiveGame } from '../entities/ActiveGame';
+import { GameDirection } from '../valueObjects/GameDirection';
+import { GameStatus } from '../valueObjects/GameStatus';
 import { ActiveGameSlotFixtures } from './ActiveGameSlotFixtures';
 import { GameCardSpecFixtures } from './GameCardSpecFixtures';
 import { GameSpecFixtures } from './GameSpecFixtures';
