+++ packages/backend/apps/game/backend-game-domain/src/games/domain/query/GameFindQuery.ts
@@ -1,4 +1,4 @@
-import { GameStatus } from '../models/GameStatus';
+import { GameStatus } from '../valueObjects/GameStatus';
 import { GameSlotFindQuery } from './GameSlotFindQuery';
 import { GameStateFindQuery } from './GameStateFindQuery';
