+++ packages/backend/apps/game/backend-game-domain/src/games/domain/valueObjects/ActiveGameState.ts
