+++ packages/backend/apps/game/backend-game-domain/src/games/domain/valueObjects/BaseGameSlot.ts
